requirements.txt
ebooklib
beautifulsoup4
httpx
argparse
python-dotenv